refactor: extract custom-rules DTOs/helpers to sibling module + fix computed-field round-trip (#2253)
